refactor: 축제 카테고리 클릭 시 축제 리스트 무한스크롤 패칭 쿼리 분리
